// TradeX MCP/Knowledge based/lotte-bridge-main/src/models/request/ICancelRightRequest.ts
import { Models } from 'tradex-common';

export interface ICancelRightRequest extends Models.IDataRequest {
  amount: number;
  bankCode: string;
  baseDate: string;
  quantity: number;
  rightType: string;
  stockCode: string;
  subNumber: string;
  bankAccount: string;
  tradeNumber: string;
  accountNumber: string;
  sequenceNumber: number;
}